+++ lib/model/controller.dart
@@ -1,5 +1,9 @@
+import 'dart:async';
+
 import 'package:mx_novice_guide/model/guide_step.dart';
 import 'package:mx_novice_guide/widget/novice_guide.dart';
+
+typedef FutureCallback = FutureOr Function();
 
 abstract class NoviceGuideController {
   static NoviceGuideController create() => NoviceGuideControllerImpl();
@@ -15,9 +19,25 @@ abstract class NoviceGuideController {
 
   Future<void> finish();
 
-  Future<void> next();
+  /// [waitCurrentEnd] - 是否等待當前步驟關閉後再跳下個步驟
+  /// [currentEndWithAnimation] - 關閉當前步驟是否有動畫
+  /// [onCurrentEndComplete] -
+  ///         當前步驟完全關閉後回調, 若需要在此執行等待操作, 可以在此回傳Future, 當[waitCurrentEnd]為true時有效
+  Future<void> next({
+    bool waitCurrentEnd = true,
+    bool currentEndWithAnimation = true,
+    FutureCallback? onCurrentEndComplete,
+  });
 
-  Future<void> previous();
+  /// [waitCurrentEnd] - 是否等待當前步驟關閉後再跳上個步驟
+  /// [currentEndWithAnimation] - 關閉當前步驟是否有動畫
+  /// [onCurrentEndComplete] -
+  ///         當前步驟完全關閉後回調, 若需要在此執行等待操作, 可以在此回傳Future, 當[waitCurrentEnd]為true時有效
+  Future<void> previous({
+    bool waitCurrentEnd = true,
+    bool currentEndWithAnimation = true,
+    FutureCallback? onCurrentEndComplete,
+  });
 
   bool canNext();
 
@@ -51,19 +71,35 @@ class NoviceGuideControllerImpl implements NoviceGuideController {
   }
 
   @override
-  Future<void> next() async {
+  Future<void> next({
+    bool waitCurrentEnd = true,
+    bool currentEndWithAnimation = true,
+    FutureCallback? onCurrentEndComplete,
+  }) async {
     if (_client == null) {
       return;
     }
-    return _client!.next();
+    return _client!.next(
+      waitCurrentEnd: waitCurrentEnd,
+      currentEndWithAnimation: currentEndWithAnimation,
+      onCurrentEndComplete: onCurrentEndComplete,
+    );
   }
 
   @override
-  Future<void> previous() async {
+  Future<void> previous({
+    bool waitCurrentEnd = true,
+    bool currentEndWithAnimation = true,
+    FutureCallback? onCurrentEndComplete,
+  }) async {
     if (_client == null) {
       return;
     }
-    return _client!.previous();
+    return _client!.previous(
+      waitCurrentEnd: waitCurrentEnd,
+      currentEndWithAnimation: currentEndWithAnimation,
+      onCurrentEndComplete: onCurrentEndComplete,
+    );
   }
 
   @override