+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## 1.0.1+1
+- MxNoviceGuide.next() and MxNoviceGuide.previous() methods add parameters
+  * waitCurrentEnd /// wait for the current step to end focus
+  * currentEndWithAnimation /// is current step end focus with animation
+  * onCurrentEndComplete /// current step end focus complete callback
+
+- Optimize focus animation
+
 ## 1.0.0
 - init version